Fix 1 - Layerweight and Override Layer removal
Fixed Exercise 2:

"A layerweight and an override layer is added to the animator when it wasn't necessary, this doesnt allow the Appear animation to play"

Removed the override and the layerweight and replaced it with a boolean parameter IsTransitionAllowed. When playing manual animations, this parameter is set to false momentarily and the manual animation that was asked to play will do so.